+++ src/main/java/svenhjol/charmony/runestones/common/features/runestones/Advancements.java
@@ -20,4 +20,16 @@ public class Advancements extends Setup<Runestones> {
     public void travelledHomeViaRunestone(Player player) {
         AdvancementHelper.trigger("travelled_home_via_runestone", player);
     }
+
+    public void learnedFiveWords(Player player) {
+        AdvancementHelper.trigger("learned_five_words", player);
+    }
+
+    public void learnedFifteenWords(Player player) {
+        AdvancementHelper.trigger("learned_fifteen_words", player);
+    }
+
+    public void learnedThirtyWords(Player player) {
+        AdvancementHelper.trigger("learned_thirty_words", player);
+    }
 }